encoders/requirements.txt
# encoders dependencies
# Vision stream encoder with CLIP

# Core
numpy>=1.24.0
opencv-python>=4.8.0

# CLIP model
torch>=2.0.0
open-clip-torch>=2.20.0
pillow>=9.5.0

# Optional: matplotlib for drift plot (can be commented out)
matplotlib>=3.7.0

